Push static layout logic to final subclass
