+++ src/main/java/jobicade/gooey/layout/Layout.java
@@ -2,77 +2,37 @@ package jobicade.gooey.layout;
 
 import com.google.common.collect.Range;
 
-import jobicade.gooey.GooeyMath;
 import jobicade.gooey.geom.Point;
 import jobicade.gooey.geom.Rect;
 
 /**
- * Tracks the position and size of a movable rectangular object.
+ * Tracks the position and size of a movable rectangular object. Code using a
+ * layout is expected to first request a size ({@link #requestSize(Point)}) then
+ * use that size to set bounds for the layout ({@link #setBounds(Rect)}).
+ *
+ * <p>Layout implementations should avoid size constraints that are likely to
+ * change between layout and rendering.
  */
-public final class Layout {
+public abstract class Layout {
     /**
-     * @return A new layout accepting a fixed size.
+     * @return A new static layout accepting a fixed size.
      */
     public static Layout fixedSize(Point size) {
-        return new Layout(size, Range.singleton(size.getX()), Range.singleton(size.getY()));
+        return new StaticLayout(size, Range.singleton(size.getX()), Range.singleton(size.getY()));
     }
 
     /**
-     * @return A new layout accepting any size with a size hint.
+     * @return A new static layout accepting any size with a size hint.
      */
     public static Layout anySize(Point sizeHint) {
-        return new Layout(sizeHint, Range.atLeast(0), Range.atLeast(0));
+        return new StaticLayout(sizeHint, Range.atLeast(0), Range.atLeast(0));
     }
 
     /**
-     * @return A new layout accepting sizes within range and with a size hint.
+     * @return A new static layout accepting sizes within range and with a size hint.
      */
     public static Layout sizeRange(Point sizeHint, Range<Integer> widthRange, Range<Integer> heightRange) {
-        return new Layout(sizeHint, widthRange, heightRange);
-    }
-
-    private final Range<Integer> widthRange;
-    private final Range<Integer> heightRange;
-
-    private Layout(Point sizeHint, Range<Integer> widthRange, Range<Integer> heightRange) {
-        if (sizeHint.getX() < 0 || sizeHint.getY() < 0) {
-            throw new IllegalArgumentException("Negative size hint");
-        }
-
-        this.sizeHint = sizeHint;
-        this.widthRange = widthRange;
-        this.heightRange = heightRange;
-
-        if (widthRange.isEmpty() || heightRange.isEmpty()) {
-            throw new IllegalArgumentException("Empty size range");
-        } else if (GooeyMath.lowerBound(widthRange) < 0 || GooeyMath.lowerBound(heightRange) < 0) {
-            throw new IllegalArgumentException("Negative size range");
-        } else if (!checkSize(sizeHint)) {
-            throw new IllegalArgumentException("Size hint out of range");
-        }
-    }
-
-    /**
-     * @return {@code true} if {@code size} is valid.
-     */
-    public boolean checkSize(Point size) {
-        return widthRange.contains(size.getX()) && heightRange.contains(size.getY());
-    }
-
-    /**
-     * @return The closest valid size to {@code size}.
-     */
-    public Point clampSize(Point size) {
-        return new Point(GooeyMath.clamp(widthRange, size.getX()), GooeyMath.clamp(heightRange, size.getY()));
-    }
-
-    private final Point sizeHint;
-    /**
-     * The size hint should be chosen as the size of the layout if there are no
-     * further external decisions or constraints.
-     */
-    public Point getSizeHint() {
-        return sizeHint;
+        return new StaticLayout(sizeHint, widthRange, heightRange);
     }
 
     private Rect bounds;
@@ -81,6 +41,8 @@ public final class Layout {
     }
 
     /**
+     * The size of the bounds is checked during this method.
+     *
      * @throws IllegalArgumentException If {@code bounds} has an invalid size.
      * @see #checkSize(Point)
      */
@@ -90,4 +52,28 @@ public final class Layout {
         }
         this.bounds = bounds;
     }
+
+    /**
+     * @return {@code true} if {@code size} is valid for this layout.
+     */
+    public boolean checkSize(Point size) {
+        return size.equals(requestSize(size));
+    }
+
+    /**
+     * Chooses a valid size in response to a requested size. If the argument
+     * is a valid size, the return value should be equal to it. For example, a
+     * layout with a maximum size of (100,100) should return (50,50) in response
+     * to (50,50) but (100,100) in response to (150,150).
+     *
+     * @return A valid size.
+     * @see #checkSize(Point)
+     */
+    public abstract Point requestSize(Point size);
+
+    /**
+     * @return A valid size that should be used for this layout given no further
+     * constraints.
+     */
+    public abstract Point getSizeHint();
 }
